Iteration 2: remove gold/commission from all calculations, fix 9 failing tests
- Remove gold_fee_fraction from scorer spread (raw spread only)
- Remove gold fees from triangular arbitrage edge weights
- Remove gold fees from recipe profit calculations
- Remove gold fee deduction from storage value net_value
- Set fee_fraction=0 and gold_fee_actual=0 in routes_arbitrage
- Fix test_momentum: expect min_volatility floor (0.01)
- Fix test_pipeline_cache: change cache_ttl to float type
- Fix test_scheduler: mock get_snapshot() instead of provider
- Fix test_daily_stats: store effective_feature_config in LightGBM
- Update test_scorer, test_recipe, test_storage_value for fee-less formulas
- All 361 tests pass, 0 failures

+++ backend/config.py
@@ -25,7 +25,7 @@ class DataConfig(BaseModel):
     primary_provider: str = "poe2scout"
     fallback_provider: str = "official"
     poe2scout_base_url: str = "https://api.poe2scout.com/api"
-    cache_ttl_prices_minutes: int = 5
+    cache_ttl_prices_minutes: float = 5
     cache_ttl_history_hours: int = 24
     cache_ttl_metadata_hours: int = 1
     rate_limit_per_second: float = 1.0
